+++ sydtest-mutation-plugin/default.nix
@@ -1,19 +1,20 @@
-{ mkDerivation, base, bytestring, containers, directory, filepath
-, ghc, ghc-boot, lib, mtl, opt-env-conf, opt-env-conf-test, path
-, path-io, safe-coloured-text, sydtest, sydtest-discover
-, sydtest-mutation-runtime, template-haskell, text
+{ mkDerivation, aeson, base, bytestring, containers, directory
+, filepath, ghc, ghc-boot, lib, mtl, opt-env-conf
+, opt-env-conf-test, path, path-io, safe-coloured-text, sydtest
+, sydtest-discover, sydtest-mutation-runtime, template-haskell
+, text
 }:
 mkDerivation {
   pname = "sydtest-mutation-plugin";
-  version = "0.2.1.0";
+  version = "0.3.0.0";
   src = ./.;
   libraryHaskellDepends = [
-    base bytestring containers directory filepath ghc ghc-boot mtl
-    opt-env-conf path path-io safe-coloured-text
+    aeson base bytestring containers directory filepath ghc ghc-boot
+    mtl opt-env-conf path path-io safe-coloured-text
     sydtest-mutation-runtime template-haskell text
   ];
   testHaskellDepends = [
-    base containers ghc opt-env-conf-test path sydtest text
+    aeson base containers ghc opt-env-conf-test path sydtest text
   ];
   testToolDepends = [ sydtest-discover ];
   homepage = "https://github.com/NorfairKing/sydtest#readme";