+++ sydtest-mutation-plugin/src/Test/Syd/Mutation/Plugin/Operator/Util.hs
@@ -12,6 +12,7 @@ module Test.Syd.Mutation.Plugin.Operator.Util
     mkConstLambda,
     arrowTy,
     prefixFormPreview,
+    unwrapWrap,
   )
 where